+++ docs/Changelog.md
@@ -9,6 +9,7 @@
 - Biker map theme improvements [#1192](https://github.com/mapsforge/vtm/issues/1192)
 - Rename `MapDatabase` to `MapFile` [#1184](https://github.com/mapsforge/vtm/pull/1184)
   - Rename `MultiMapDatabase` to `MultiMapFile`
+- Update MapBox vector tile dependency
 - Minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.25.0)